chore: split OpenCodeClientTests by domain
Move suites out of the 4700-line monolith into Models, Connection,
SessionFlow, CarMode, AIBuilderVoice, Rendering, and TestDoubles.
No assertion changes.
